Python/w/requirements.txt
absl-py==1.3.0
asgiref==3.5.2
asttokens==2.0.8
astunparse==1.6.3
backcall==0.2.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cloudpickle==2.2.0
colorama==0.4.6
contourpy==1.0.5
crispy-bootstrap5==0.7
cryptography==38.0.1
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==4.1
django-allauth==0.51.0
django-appconf==1.0.5
django-classy-tags==3.0.1
django-compressor==4.1
django-crispy-forms==1.14.0
django-sekizai==4.0.0
djangorestframework==3.14.0
executing==1.1.1
filelock==3.8.0
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
gym==0.26.2
gym-notices==0.0.8
h5py==3.7.0
idna==3.4
ipython==8.5.0
jedi==0.18.1
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-rl==0.4.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
matplotlib-inline==0.1.6
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
pipenv==2022.10.12
platformdirs==2.5.2
prompt-toolkit==3.0.31
protobuf==3.19.6
pulpo-forms==1.1
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.5
rcssmin==1.1.0
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.1
rjsmin==1.2.0
rsa==4.9
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
traitlets==5.5.0
typing_extensions==4.4.0
tzdata==2022.5
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
